+++ include/base_types.h
@@ -51,7 +51,7 @@ typedef uint64_t   UINT64;
 
 /* and the ever-so-important array size macro */
 #ifndef ARRAY_SIZE
-#define ARRAY_SIZE(x)    (sizeof(x) / sizeof((x)[0]))
+#define ARRAY_SIZE(x)    (size_t)(sizeof(x) / sizeof((x)[0]))
 #endif
 
 #endif /* BASE_TYPES_H_INCLUDED */
